[TS-286] Refactor: StarCard, StarCardDetail 폴더를 StarPage 폴더 내부로 이동
Star 관련 페이지(메인, 카드, 카드 상세)를 하나로 묶는게 보기 편할 것 같아서 StarPage 폴더 내부로 이동함

별자리 메인 페이지까지 추가되면 아래와 같은 폴더 구조가 될 예정
StarPage/StarMain
StarPage/StarCard
StarPage/StarCardDetail
