--- src/components/StarCardDetail/Img/PossessionLabel.style.ts
@@ -1,13 +1,0 @@
-import { css } from "@emotion/react";
-
-export const possessionLabelStyle = css`
-	width: 100px;
-	padding: 16px 0;
-	border-radius: 15px 0 15px 0;
-	position: absolute;
-	top: 0;
-	font-weight: 700;
-	text-align: center;
-	color: #fff;
-	background-color: #385678;
-`;
