--- src/components/StarCardDetail/Img/PossessionLabel.tsx
@@ -1,5 +1,0 @@
-import { possessionLabelStyle } from "@/components/StarCardDetail/Img/PossessionLabel.style";
-
-export default function PossessionLabel() {
-	return <div css={possessionLabelStyle}>보유</div>;
-}
