+++ src/components/StarPage/StarCard/CategoryTab.tsx
