+++ src/components/StarPage/StarCard/StarCard.style.ts
