+++ src/components/StarPage/StarCard/StarCard.tsx
@@ -7,7 +7,7 @@ import {
 	haveLabelStyle,
 	titleStyle,
 	selectedStyle,
-} from "@/components/StarCard/StarCard.style";
+} from "@/components/StarPage/StarCard/StarCard.style";
 
 // TODO: API 연결 후 삭제 or 수정 예정
 interface StarCardProps {
