+++ src/components/StarPage/StarCardDetail/CategoryLabel.style.ts
