+++ src/components/StarPage/StarCardDetail/CategoryLabel.tsx
@@ -1,6 +1,6 @@
-import Label from "@/components/StarCardDetail/Label";
+import Label from "@/components/StarPage/StarCardDetail/Label";
 
-import { categoryLabelStyle } from "@/components/StarCardDetail/CategoryLabel.style";
+import { categoryLabelStyle } from "@/components/StarPage/StarCardDetail/CategoryLabel.style";
 
 // TODO: API 연결 후 삭제 예정
 const data = [
