+++ src/components/StarPage/StarCardDetail/Img/Img.style.ts
