+++ src/components/StarPage/StarCardDetail/Img/Star.style.ts
