+++ src/components/StarPage/StarCardDetail/Img/Star.tsx
@@ -1,6 +1,6 @@
 import starImg from "@/assets/image/card-detail-star.png";
 
-import { starImgStyle } from "@/components/StarCardDetail/Img/Star.style";
+import { starImgStyle } from "@/components/StarPage/StarCardDetail/Img/Star.style";
 
 export default function Star({ starNumber }: { starNumber: number }) {
 	return (
