+++ src/components/StarPage/StarCardDetail/Label.style.ts
