+++ src/components/StarPage/StarCardDetail/Story.style.ts
