+++ src/components/StarPage/StarCardDetail/Story.tsx
