+++ src/components/StarPage/StarCardDetail/Title.style.ts
